Eliminated the old broadcast integration test, replaced with the new
